Do not move if not exists

+++ scripts/editors.sh
@@ -4,11 +4,17 @@ set -u
 
 if command -v git &>/dev/null; then
   # emacs
-  mv "$HOME/.emacs" "$HOME/.emacs.bak"
-  mv "$HOME/.emacs.d" "$HOME/.emacs.d.bak"
+  if [ -e "$HOME/.emacs" ]; then
+    mv "$HOME/.emacs" "$HOME/.emacs.bak"
+  fi
+  if [ -e "$HOME/.emacs.d" ]; then
+    mv "$HOME/.emacs.d" "$HOME/.emacs.d.bak"
+  fi
   git clone "https://github.com/syl20bnr/spacemacs" "$HOME/.emacs.d"
 
   # ViM
-  mv "$HOME/.vim/bundle/Vundle.vim" "$HOME/.vim/bundle/Vundle.vim.bak"
+  if [ -e "$HOME/.vim/bundle/Vundle.vim" ]; then
+    mv "$HOME/.vim/bundle/Vundle.vim" "$HOME/.vim/bundle/Vundle.vim.bak"
+  fi
   git clone "https://github.com/gmarik/Vundle.vim.git" "$HOME/.vim/bundle/Vundle.vim"
 fi
